Overhaul
- Changed the interface so that it does not require custom classes
implementing an interface, but rather, can be used with base types.  The
interface is now closer to that of IDictionary<TKey, TValue>
- Incorporated refactorings and fixes from other forks of the project
- Incorporated unit tests from other similar projects
- Changed usage of comparers so that the implementation supports
specifying custom comparers for pre-existing types (e.g. strings can be
used as keys and can leverage any of the standard .Net string comparers)
- Removed the Async code for now (not required for my use case, and
therefore, not worth fixing for the new interface).